Don't require a mutable borrow for `find_widgets()`
This function previously required a mutable borrow of `self`
(borrowing the widget instance itself) in order to traverse a path
and return a reference to a particular widget.
This was only done to update the cache of found widgets,
in order to accelerate future searches for the same path ID.

This is not necessary, and in fact it results in Makepad app devs
having to adopt unnecessary clones and other inefficient code patterns
in more complex app code.