Only trigger actions for most widgets upon a primary mouse button/touch hit
Many widgets that can be clicked have previously treated *any* click
as if it were a primary mouse button click (or touch).
This has led to unexpected results, such as activating a button
upon a right click or a back/forward button click, which is generally
non-standard behavior.

This changeset ensures that many widget actions only occur if a
regular touch press or a primary mouse button click occurs.
To make this easy, we add the `is_primary_hit()` function
which can be conveniently added to a branch when matching on `hits()`:
```rust
match event.hits(...) {
    Hit::FingerDown(fe) if fe.is_primary_hit() => { ... }
    Hit::FingerUp(fe) if fe.is_primary_hit() => { ... }
}
```