WIP saving work on modifying View's event.hits handler to not mark any events as handled.
Change the `find_digit_for_captured_area()` conditional to allow
a `FingerDown` hit to be returned if the digit ID is the same.
This is an important change that allows a `FingerDown` hit event
to be handled multiple times.
However, we're not yet 100% sure if that works in all cases.